requirements.txt
flask
flask-bootstrap4
wordcloud